Provide a codefix for legacy service usage and add tests
The codefix should be helpful in migrating larger codebases that used the previous approach (since we allow batch fixing in an entire solution).

The added project also means we need a separate packaging project (can't reference the codefix project from the analyzer one because that can cause the analyzer to not load in the reduced scope of the compiler). The codefix project is only intended by consumption by an editor that can apply fixes.

We now have a standalone packaging project that contains the eula and readme plus references to the other two.